requirements.txt
# Core dependencies
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3

# Geospatial
folium
streamlit-folium==0.16.0
geopandas==0.14.2
shapely==2.0.2
rasterio==1.3.9

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3

# Data fetching
requests==2.31.0
Pillow==10.2.0

# Optional: For advanced satellite analysis
# torch==2.1.2
# torchvision==0.16.2
# segmentation-models-pytorch==0.3.3
# sentinelhub==3.9.1

# Optional: For raster data processing
# earthpy==0.9.4
# rasterstats==0.19.0